+++ ncpeh-simulation-mock/src/main/java/de/gematik/ncpeh/api/mock/NCPeHMockApiImpl.java
@@ -1,5 +1,5 @@
 /*
- * Copyright (c) 2024-2025 gematik GmbH
+ * Copyright 2024-2025 gematik GmbH
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -12,13 +12,20 @@
  * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
  * See the License for the specific language governing permissions and
  * limitations under the License.
+ *
+ * ******
+ *
+ * For additional notes and disclaimer from gematik and in case of changes by gematik find details in the "Readme" file.
  */
 
 package de.gematik.ncpeh.api.mock;
 
+import com.fasterxml.jackson.databind.ObjectMapper;
 import de.gematik.ncpeh.api.NcpehSimulatorApi;
 import de.gematik.ncpeh.api.mock.builder.HttpMessageFactory;
+import de.gematik.ncpeh.api.mock.builder.RetrieveDocumentMessagesBuilder;
 import de.gematik.ncpeh.api.mock.builder.SimulatorCommunicationDataBuilder;
+import de.gematik.ncpeh.api.mock.data.Patient;
 import de.gematik.ncpeh.api.request.FindDocumentsRequest;
 import de.gematik.ncpeh.api.request.IdentifyPatientRequest;
 import de.gematik.ncpeh.api.request.ProvideAndRegisterSetOfDocumentsRequest;
@@ -29,18 +36,23 @@ import jakarta.ws.rs.core.HttpHeaders;
 import jakarta.ws.rs.core.MediaType;
 import jakarta.ws.rs.core.Response;
 import jakarta.ws.rs.core.Response.ResponseBuilder;
+import java.io.IOException;
+import java.util.Base64;
 import java.util.Optional;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
 
 /**
  * Implementation of the NCPeH Trigger Interface API It's all Spring and Apache CXF magic, so pretty
  * slim implementation
  */
+@Slf4j
 @Service
 public class NCPeHMockApiImpl implements NcpehSimulatorApi {
 
   public static final String FILE_EXTENSIONS_XML = ".xml";
   public static final String HEADER_X_NCPEH_MOCK_RESPONSE = "X-NCPeHMock-Response";
+  public static final String HEADER_X_NCPEH_MOCK_PATIENT = "X-NCPeHMock-Patient";
   @Context private HttpHeaders headers;
 
   public static final String oidAssigningAuthority_psa = "1.2.276.0.76.4.298";
@@ -48,6 +60,12 @@ public class NCPeHMockApiImpl implements NcpehSimulatorApi {
 
   public static final String XDS_DOCUMENT_ENTRY_CLASS_CODE_EPED =
       "('57833-6^^2.16.840.1.113883.6.1')";
+
+  private final ObjectMapper mapper;
+
+  public NCPeHMockApiImpl(final ObjectMapper mapper) {
+    this.mapper = mapper;
+  }
 
   @Override
   public Response identifyPatient(final IdentifyPatientRequest request) {
@@ -96,13 +114,17 @@ public class NCPeHMockApiImpl implements NcpehSimulatorApi {
   @Override
   public Response retrieveDocument(final RetrieveDocumentRequest request) {
     final var fileName = getFileNameFromRequestHeader();
+    final var patient = getPatientFromRequestHeader();
 
     return okResponseBuilder()
         .entity(
             SimulatorCommunicationDataBuilder.newInstance()
                 .requestMessage(HttpMessageFactory.buildRetrieveDocumentRequest(request))
                 .responseMessage(
-                    HttpMessageFactory.buildRetrieveDocumentResponse(request, fileName))
+                    HttpMessageFactory.buildRetrieveDocumentResponse(
+                        RetrieveDocumentMessagesBuilder.buildFromRequestAndPatient(
+                            request, patient),
+                        fileName))
                 .build())
         .build();
   }
@@ -110,13 +132,17 @@ public class NCPeHMockApiImpl implements NcpehSimulatorApi {
   @Override
   public Response retrieveSetOfDocuments(final RetrieveSetOfDocumentsRequest request) {
     final var fileName = getFileNameFromRequestHeader();
+    final var patient = getPatientFromRequestHeader();
 
     return okResponseBuilder()
         .entity(
             SimulatorCommunicationDataBuilder.newInstance()
                 .requestMessage(HttpMessageFactory.buildStandardRetrieveSetOfDocumentsRequest())
                 .responseMessage(
-                    HttpMessageFactory.buildStandardRetrieveSetOfDocumentsResponse(fileName))
+                    HttpMessageFactory.buildRetrieveDocumentResponse(
+                        RetrieveDocumentMessagesBuilder.buildFromRequestAndPatient(
+                            request, patient),
+                        fileName))
                 .build())
         .build();
   }
@@ -145,6 +171,21 @@ public class NCPeHMockApiImpl implements NcpehSimulatorApi {
   private String getFileNameFromRequestHeader() {
     return Optional.ofNullable(headers.getHeaderString(HEADER_X_NCPEH_MOCK_RESPONSE))
         .map(name -> name + FILE_EXTENSIONS_XML)
+        .orElse(null);
+  }
+
+  private Patient getPatientFromRequestHeader() {
+    return Optional.ofNullable(headers.getHeaderString(HEADER_X_NCPEH_MOCK_PATIENT))
+        .map(
+            base64json -> {
+              try {
+                return mapper.readValue(Base64.getDecoder().decode(base64json), Patient.class);
+              } catch (final IOException e) {
+                // nothing to do
+                log.warn("Failed to parse patient from header", e);
+              }
+              return null;
+            })
         .orElse(null);
   }
 }